pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "phantom"
version = "2.0.0"
description = "AI-Powered Document Intelligence & Classification Pipeline"
readme = "docs/README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "kernelcore", email = "kernelcore@voidnix.dev" }
]
keywords = [
    "ai",
    "llm",
    "rag",
    "document-processing",
    "nlp",
    "embeddings",
    "classification",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "rich>=13.0",
    "typer>=0.9",
    
    # Data processing
    "pandas>=2.0",
    "numpy>=1.24",
    "polars>=0.20",
    
    # NLP & ML
    "sentence-transformers>=2.2",
    "transformers>=4.30",
    "tiktoken>=0.5",
    "nltk>=3.8",
    "scikit-learn>=1.3",
    
    # Embeddings & Vector DB
    "faiss-cpu>=1.7",
    "chromadb>=0.4",
    
    # HTTP & API
    "requests>=2.31",
    "httpx>=0.24",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    
    # Utilities
    "python-magic>=0.4",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.5",
    "pre-commit>=3.4",
]
gpu = [
    "torch>=2.0",
    "faiss-gpu>=1.7",
]
all = [
    "phantom[dev]",
    "phantom[gpu]",
]

[project.scripts]
phantom = "phantom.cli.main:app"
phantom-api = "phantom.api.app:serve"

[project.urls]
Homepage = "https://github.com/kernelcore/phantom"
Documentation = "https://github.com/kernelcore/phantom/docs"
Repository = "https://github.com/kernelcore/phantom"

[tool.hatch.build.targets.wheel]
packages = ["src/phantom"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
